+++ Changelog.md
@@ -17,6 +17,7 @@ v2.4x.0 (not published yet)
 v2.38.6 (not published yet)
 ====================
 
+- Log: Extend logging for Dolby Vision metadata files
 - UI: Show TaskDialog timeout when set
 - UI: Fix AutoCrop not working properly on videos with multiple aspect ratios ([#1324](/../../issues/1324))